Deal with "every trip day is the first" bug in AIMS.
The trip day on multi-day trip sheet is no longer incremented between the
duties. This made the date incorrect for sectors on the second day and
later. Now assume that each duty in a multi duty block starts on incremental
days. Maybe switch to reading the 29Apr2020 code at some point.