Fixed brief roster parsing unit test
Duties such as standby and SEP training are ambiguous in the AIMS brief
roster. They sometimes appear as independent duties and sometimes as a sector in
a trip, and sometimes they appear as both at the same time. For consistency,
when one of these duties appears as an independent duty, a single QUASI sector
of the type that would be found if it was part of a trip is added to that duty.